monitor/setup_alerts.py
import logging
from azure.monitor.opentelemetry import configure_azure_monitor

# Configure OpenTelemetry to export to Azure Monitor
# Replace the connection string with your Azure Application Insights connection string
# This is a key step for monitoring in production.
try:
    configure_azure_monitor(
        connection_string="<YOUR_APP_INSIGHTS_CONNECTION_STRING>"
    )
    logging.info("Azure Monitor configured successfully.")
except Exception as e:
    logging.error(f"Failed to configure Azure Monitor: {str(e)}")
    
def setup_metric_alerts():
    """
    A placeholder function to show how you would set up alerts.
    In a real-world scenario, this would be done using Terraform or other IaC tools.
    """
    logging.info("Setting up metric alerts in Azure Monitor...")
    # This function would use Azure SDKs to create alert rules.
    # For example, an alert for high gas prices or a drop in data freshness.
    
    # Example alert logic (pseudo-code):
    #
    # if high_gas_price:
    #     send_notification("Gas prices are high: > 100 Gwei")
    #
    # if data_freshness_check_failed:
    #     send_notification("Data freshness check failed.")
    
    logging.info("Metric alerts setup complete (pseudo-code).")
