requirements.txt
# =============================================================================
# Core Framework & Azure SDK
# =============================================================================
aiohttp==3.9.5
fastapi==0.104.1
uvicorn==0.23.2
azure-eventhub==5.11.4
azure-storage-blob==12.19.0
azure-storage-file-datalake==12.14.0
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0
azure-monitor-opentelemetry==1.2.0
azure-kusto-data==4.3.1
azure-kusto-ingest==4.3.1
requests==2.31.0
httpx==0.25.2

# =============================================================================
# Data Processing & Analytics
# =============================================================================
#pyspark==3.5.0
delta-spark==2.4.0
pyarrow==14.0.1
pandas==2.3.0
numpy==2.0.0

# =============================================================================
# Data Validation
# =============================================================================
pydantic==2.5.0
jsonschema==4.20.0

# =============================================================================
# Monitoring & Observability
# =============================================================================
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
structlog==23.2.0
sentry-sdk==1.38.0

# =============================================================================
# Ethereum & Blockchain
# =============================================================================
web3==6.12.0
eth-account==0.9.0
eth-hash==0.5.2

# =============================================================================
# Utilities
# =============================================================================
python-dotenv==1.0.0
pyyaml==6.0.1
python-dateutil==2.8.2
pytz==2023.3
tenacity==8.2.3